refactor: reorganize exports & re group features
