+++ src/index.ts
@@ -1,7 +1,6 @@
-export * from './misc'
-export * from './units/formatter'
-export * from './units/multiple'
-export * from './units/percent'
-export * from './units/price'
-export * from './units/scale'
-export * from './units/unscale'
+export * from './arithmetic/index'
+export * from './core/index'
+export * from './formatting/index'
+export * from './scaling/index'
+export * from './units/index'
+export * from './utils/index'
