+++ src/scaling/index.ts
@@ -1,0 +1,3 @@
+export * from './old-scale'
+export * from './rescale'
+export * from './scale'
